+++ controllers/operator/mongodbstandalone_controller.go
@@ -129,7 +129,7 @@ func (r *ReconcileMongoDbStandalone) Reconcile(_ context.Context, request reconc
 	log.Infow("Standalone.Spec", "spec", s.Spec)
 	log.Infow("Standalone.Status", "status", s.Status)
 
-	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, s)
+	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, s, log)
 	if err != nil {
 		return r.updateStatus(s, workflow.Failed(err.Error()), log)
 	}
@@ -268,7 +268,7 @@ func (r *ReconcileMongoDbStandalone) OnDelete(obj runtime.Object, log *zap.Sugar
 
 	log.Infow("Removing standalone from Ops Manager", "config", s.Spec)
 
-	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, s)
+	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, s, log)
 	if err != nil {
 		return err
 	}