+++ controllers/operator/mongodbuser_controller.go
@@ -128,7 +128,7 @@ func (r *MongoDBUserReconciler) Reconcile(_ context.Context, request reconcile.R
 		return reconcile.Result{}, nil
 	}
 
-	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, mdb)
+	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, mdb, log)
 	if err != nil {
 		return r.updateStatus(user, workflow.Failed(err.Error()), log)
 	}
@@ -153,7 +153,7 @@ func (r *MongoDBUserReconciler) delete(obj interface{}, log *zap.SugaredLogger) 
 		return err
 	}
 
-	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, mdb)
+	projectConfig, credsConfig, err := project.ReadConfigAndCredentials(r.client, mdb, log)
 	if err != nil {
 		return err
 	}