+++ RELEASE_NOTES.md
@@ -4,6 +4,7 @@
 
 ## Kubernetes Operator
 * When deploying to multiple namespaces, imagePullSecrets has to be created only in the namespace where the Operator is installed. From here, the Operator will be sync this secret across all watched namespaces.
+* The credentials secret used by the Operator now accepts the pair of fields `publicKey` and `privateKey`. These should be preferred to the existent `user` and `publicApiKey` when using Programmatic API Keys in Ops Manager.
 
 ## MongoDB Resource
 * Bug fixes